Revert "http-nats-wait"
