+++ core/build.gradle
@@ -10,5 +10,4 @@ dependencies {
 	compile ('com.fasterxml.jackson.core:jackson-core:2.7.5') {force=true}
 	compile 'com.jayway.jsonpath:json-path:2.2.0'
 	compile 'dnsjava:dnsjava:2.1.8'
-	compile 'net.thisptr:jackson-jq:+'
 }
