+++ core/src/main/java/com/netflix/conductor/core/events/ScriptEvaluator.java
@@ -18,21 +18,10 @@
  */
 package com.netflix.conductor.core.events;
 
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.common.cache.CacheBuilder;
-import com.google.common.cache.CacheLoader;
-import com.google.common.cache.LoadingCache;
-import net.thisptr.jackson.jq.JsonQuery;
-import net.thisptr.jackson.jq.exception.JsonQueryException;
-
-import javax.annotation.Nonnull;
 import javax.script.Bindings;
 import javax.script.ScriptEngine;
 import javax.script.ScriptEngineManager;
 import javax.script.ScriptException;
-import java.util.List;
-import java.util.concurrent.TimeUnit;
 
 /**
  * @author Viren
@@ -41,9 +30,7 @@ import java.util.concurrent.TimeUnit;
 public class ScriptEvaluator {
 
 	private static ScriptEngine engine = new ScriptEngineManager().getEngineByName("nashorn");
-	private static LoadingCache<String, JsonQuery> queryCache = createQueryCache();
-	private static final ObjectMapper om = new ObjectMapper();
-
+	
 	private ScriptEvaluator(){
 		
 	}
@@ -64,25 +51,5 @@ public class ScriptEvaluator {
 		bindings.put("$", input);
 		return engine.eval(script, bindings);
 		
-	}
-
-	static String evalJq(String expression, Object payload) throws Exception {
-		JsonNode input = om.valueToTree(payload);
-		JsonQuery query = queryCache.get(expression);
-		List<JsonNode> result = query.apply(input);
-		if (result == null || result.isEmpty()) {
-			return null;
-		} else {
-			return result.get(0).asText();
-		}
-	}
-
-	private static LoadingCache<String, JsonQuery> createQueryCache() {
-		CacheLoader<String, JsonQuery> loader = new CacheLoader<String, JsonQuery>() {
-			public JsonQuery load(@Nonnull String query) throws JsonQueryException {
-				return JsonQuery.compile(query);
-			}
-		};
-		return CacheBuilder.newBuilder().expireAfterWrite(1, TimeUnit.HOURS).maximumSize(1000).build(loader);
 	}
 }
