+++ server/src/main/java/com/netflix/conductor/server/ServerModule.java
@@ -20,8 +20,11 @@ package com.netflix.conductor.server;
 
 import com.google.inject.AbstractModule;
 import com.google.inject.Provides;
-import com.netflix.conductor.contribs.HttpModule;
 import com.netflix.conductor.contribs.auth.AuthTask;
+import com.netflix.conductor.contribs.http.HttpTask;
+import com.netflix.conductor.contribs.http.HttpWaitTask;
+import com.netflix.conductor.contribs.http.RestClientManager;
+import com.netflix.conductor.contribs.http.RestClientManagerHttpWait;
 import com.netflix.conductor.contribs.json.JsonJqTransform;
 import com.netflix.conductor.contribs.validation.ValidationTask;
 import com.netflix.conductor.core.config.Configuration;
@@ -108,10 +111,11 @@ public class ServerModule extends AbstractModule {
 
 		install(new CoreModule());
 		install(new JerseyModule());
-		install(new HttpModule());
+		new AuthTask(config);
+		new HttpTask(new RestClientManager(), config);
+		new HttpWaitTask(new RestClientManagerHttpWait(), config);
 		new JsonJqTransform();
 		new ValidationTask();
-		new AuthTask(config);
 		List<AbstractModule> additionalModules = config.getAdditionalModules();
 		if (additionalModules != null) {
 			for (AbstractModule additionalModule : additionalModules) {
